app/code-interpreter/requirements.txt
slackagents==0.0.1
serpapi
python-dotenv